Add a native PC port (macOS and Windows)
Runs the decompiled C directly against SDL2 — no emulator, no ROM at run
time. platform/ reimplements the GBA hardware the game talks to: PPU and
sprite compositing, DMA, the sound FIFO and MIDI DSP, key registers, BIOS
calls, and the affine parameter manager.

Notable pieces beyond the hardware layer:

* platform/beatscript_vm.c translates func_0800cb28, the beatscript command
  processor, from its 2610-line Thumb listing. 45 of the 179 jump-table
  entries are implemented — every opcode the shipped scripts use.

* tools/bs2c.py compiles the .bs scene scripts to C, so scenes run their real
  scripts. Previously the PC build had hand-written approximations, which is
  why Karate Man's consecutive-cue patterns were missing and drifted against
  the music. Macros are resolved by following each file's own .include chain,
  and any line that survives expansion unrecognised is reported rather than
  silently dropped.

* tools/bin2s.py replaces devkitPro's bin2s (verified to produce identical
  objects), so the PC port builds with no devkitPro install.

Portability fixes the port needed: POSIX-only debug helpers guarded, SIGBUS
guarded, and auto_stubs' weak symbols made link-order-resolved on PE/COFF,
which has no usable weak definitions.

Several latent 64-bit bugs surfaced and are fixed here: pointers truncated
through s32 fields (sprite data, scene transition variables, engine event
parameters), functions relying on the GBA's r0 return idiom, and stub globals
declared smaller than their extern types.

+++ src/lib_0804ca80.h
@@ -253,7 +253,9 @@ extern void sprite_set_affine_params(struct SpriteHandler *, s16 id, s32 affineI
 extern void func_0804dc8c(struct SpriteHandler *, s16 id, u32 args); // Set Affine & Double Size Flags
 extern void sprite_set_anim_speed(struct SpriteHandler *, s16 id, u8_8 speed);
 extern u32  sprite_get_cel_dimensions(u16 *cel, u32 requestedDataType);
-extern s32  sprite_get_data(struct SpriteHandler *, s16 id, u32 requestedDataType);
+// Returns intptr_t: pointer-valued data types (animation, callback, origins)
+// must survive the round-trip on 64-bit hosts.
+extern intptr_t sprite_get_data(struct SpriteHandler *, s16 id, u32 requestedDataType);
 extern void sprite_set_callback_cel(struct SpriteHandler *, s16 id, s8 cel);
 extern void sprite_run_callback_every_cel(struct SpriteHandler *, s16 id);
 extern void sprite_id_set_data(struct SpriteHandler *, u16 memID, u32 targetDataType, u32 arg);
